requirements.txt
# Databricks SDK and API
databricks-sdk==0.48.0

# ML Libraries
mlflow>=2.11.0
scikit-learn>=1.4.0
numpy>=1.24.0
pandas>=2.0.0

# LangChain and Prompt Engineering
langchain==0.3.0
langchain-openai==0.2.0
langgraph==0.2.22
langgraph-checkpoint==1.0.11

# Data Generation
faker

# Web UI
streamlit>=1.35.0

# Utilities
python-dotenv>=1.0.0